feat: add Aside
An aside makes it clear that content is optional. Sometimes I want to
explain something in a lot of detail, but don't consider it core to what
I want to describe. An aside make it easy to skip, and makes it clear
which content I think is skip-able.

It can also have the opposite effect, and tell people who want more
details where to find them.